Preserve the tests and code, eliminate the SimplePostTool
