+++ solr/core/src/java/org/apache/solr/cli/PostTool.java
@@ -16,15 +16,149 @@
  */
 package org.apache.solr.cli;
 
+import static java.nio.charset.StandardCharsets.US_ASCII;
+import static java.nio.charset.StandardCharsets.UTF_8;
+
+import java.io.BufferedReader;
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.File;
+import java.io.FileFilter;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.io.PrintStream;
+import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.ProtocolException;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.net.URL;
+import java.net.URLEncoder;
+import java.nio.ByteBuffer;
+import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.InvalidPathException;
+import java.security.GeneralSecurityException;
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Base64;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.LinkedHashSet;
 import java.util.List;
+import java.util.Locale;
+import java.util.Map;
+import java.util.Set;
+import java.util.TimeZone;
+import java.util.regex.Pattern;
+import java.util.regex.PatternSyntaxException;
+import java.util.zip.GZIPInputStream;
+import java.util.zip.Inflater;
+import java.util.zip.InflaterInputStream;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.parsers.ParserConfigurationException;
+import javax.xml.xpath.XPath;
+import javax.xml.xpath.XPathConstants;
+import javax.xml.xpath.XPathExpression;
+import javax.xml.xpath.XPathExpressionException;
+import javax.xml.xpath.XPathFactory;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Option;
+import org.apache.solr.client.api.util.SolrVersion;
+import org.apache.solr.common.util.Utils;
+import org.apache.solr.util.RTimer;
+import org.w3c.dom.Document;
+import org.w3c.dom.Node;
+import org.w3c.dom.NodeList;
+import org.xml.sax.SAXException;
 
 /** Supports post command in the bin/solr script. */
 public class PostTool extends ToolBase {
+
+  public static final String DEFAULT_FILE_TYPES =
+      "xml,json,jsonl,csv,pdf,doc,docx,ppt,pptx,xls,xlsx,odt,odp,ods,ott,otp,ots,rtf,htm,html,txt,log";
+  static final String DATA_MODE_FILES = "files";
+  static final String DATA_MODE_ARGS = "args";
+  static final String DATA_MODE_STDIN = "stdin";
+  static final String DEFAULT_DATA_MODE = DATA_MODE_FILES;
+  static final String FORMAT_SOLR = "solr";
+  static final String DATA_MODE_WEB = "web";
+
+  private static final String DEFAULT_COMMIT = "yes";
+  private static final String DEFAULT_OPTIMIZE = "no";
+  private static final String DEFAULT_OUT = "no";
+  private static final String DEFAULT_AUTO = "no";
+  private static final String DEFAULT_RECURSIVE = "0";
+  private static final int DEFAULT_WEB_DELAY = 10;
+  private static final int MAX_WEB_DEPTH = 10;
+  public static final String DEFAULT_CONTENT_TYPE = "application/json";
+
+  // Input args
+  boolean auto = false;
+  int recursive = 0;
+  int delay = 0;
+  String fileTypes = PostTool.DEFAULT_FILE_TYPES;
+  URL solrUrl;
+  String credentials;
+  OutputStream out = null;
+  String type;
+  String format;
+  String mode = DEFAULT_DATA_MODE;
+  boolean commit;
+  boolean optimize;
+  boolean dryRun;
+
+  String[] args;
+
+  private int currentDepth;
+
+  static HashMap<String, String> mimeMap;
+  FileFilter fileFilter;
+  // Backlog for crawling
+  List<LinkedHashSet<URI>> backlog = new ArrayList<>();
+  Set<URI> visited = new HashSet<>();
+
+  static final Set<String> DATA_MODES = new HashSet<>();
+
+  // Used in tests to avoid doing actual network traffic
+  boolean mockMode = false;
+  PostTool.PageFetcher pageFetcher;
+
+  static {
+    DATA_MODES.add(PostTool.DATA_MODE_FILES);
+    DATA_MODES.add(DATA_MODE_ARGS);
+    DATA_MODES.add(DATA_MODE_STDIN);
+    DATA_MODES.add(PostTool.DATA_MODE_WEB);
+
+    mimeMap = new HashMap<>();
+    mimeMap.put("xml", "application/xml");
+    mimeMap.put("csv", "text/csv");
+    mimeMap.put("json", "application/json");
+    mimeMap.put("jsonl", "application/jsonl");
+    mimeMap.put("pdf", "application/pdf");
+    mimeMap.put("rtf", "text/rtf");
+    mimeMap.put("html", "text/html");
+    mimeMap.put("htm", "text/html");
+    mimeMap.put("doc", "application/msword");
+    mimeMap.put("docx", "application/vnd.openxmlformats-officedocument.wordprocessingml.document");
+    mimeMap.put("ppt", "application/vnd.ms-powerpoint");
+    mimeMap.put(
+        "pptx", "application/vnd.openxmlformats-officedocument.presentationml.presentation");
+    mimeMap.put("xls", "application/vnd.ms-excel");
+    mimeMap.put("xlsx", "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet");
+    mimeMap.put("odt", "application/vnd.oasis.opendocument.text");
+    mimeMap.put("ott", "application/vnd.oasis.opendocument.text");
+    mimeMap.put("odp", "application/vnd.oasis.opendocument.presentation");
+    mimeMap.put("otp", "application/vnd.oasis.opendocument.presentation");
+    mimeMap.put("ods", "application/vnd.oasis.opendocument.spreadsheet");
+    mimeMap.put("ots", "application/vnd.oasis.opendocument.spreadsheet");
+    mimeMap.put("txt", "text/plain");
+    mimeMap.put("log", "text/plain");
+  }
 
   public PostTool() {
     this(CLIO.getOutStream());
@@ -44,6 +178,7 @@ public class PostTool extends ToolBase {
     return List.of(
         Option.builder("url")
             .argName("url")
+            .longOpt("solrUrl")
             .hasArg()
             .required(false)
             .desc("<base Solr update URL>")
@@ -82,6 +217,11 @@ public class PostTool extends ToolBase {
             .desc(
                 "If recursive then delay will be the wait time between posts.  default: 10 for web, 0 for files")
             .build(),
+        Option.builder("auto")
+            .longOpt("auto")
+            .required(false)
+            .desc("If true, we'll guess type and add resourcename/url. Defaults to false.")
+            .build(),
         Option.builder("type")
             .argName("content-type")
             .hasArg(true)
@@ -92,7 +232,7 @@ public class PostTool extends ToolBase {
             .argName("<type>[,<type>,...]")
             .hasArg(true)
             .required(false)
-            .desc("default: " + SimplePostTool.DEFAULT_FILE_TYPES)
+            .desc("default: " + DEFAULT_FILE_TYPES)
             .build(),
         Option.builder("params")
             .argName("<key>=<value>[&<key>=<value>...]")
@@ -109,6 +249,11 @@ public class PostTool extends ToolBase {
             .desc(
                 "sends application/json content as Solr commands to /update instead of /update/json/docs.")
             .build(),
+        Option.builder()
+            .longOpt("dryrun")
+            .required(false)
+            .desc("Performs a dry run of the posting process without sending documents to Solr.")
+            .build(),
         SolrCLI.OPTION_CREDENTIALS);
   }
 
@@ -116,7 +261,7 @@ public class PostTool extends ToolBase {
   public void runImpl(CommandLine cli) throws Exception {
     SolrCLI.raiseLogLevelUnlessVerbose(cli);
 
-    URL solrUrl = null;
+    solrUrl = null;
     if (cli.hasOption("url")) {
       String url = cli.getOptionValue("url");
       solrUrl = new URL(url);
@@ -128,53 +273,1063 @@ public class PostTool extends ToolBase {
           "Must specify either -url or -c parameter to post documents.");
     }
 
-    String mode = SimplePostTool.DEFAULT_DATA_MODE;
+    dryRun = false;
+    if (cli.hasOption("dryrun")) {
+      dryRun = true;
+    }
+
     if (cli.hasOption("mode")) {
       mode = cli.getOptionValue("mode");
     }
-    boolean auto = true;
-    String type = null;
+    if (cli.hasOption("auto")) {
+      auto = true;
+    }
     if (cli.hasOption("type")) {
       type = cli.getOptionValue("type");
     }
-    String format =
-        cli.hasOption("format")
-            ? SimplePostTool.FORMAT_SOLR
-            : ""; // i.e not solr formatted json commands
+    format = cli.hasOption("format") ? FORMAT_SOLR : ""; // i.e not solr formatted json commands
 
-    String fileTypes = SimplePostTool.DEFAULT_FILE_TYPES;
     if (cli.hasOption("filetypes")) {
       fileTypes = cli.getOptionValue("filetypes");
     }
 
-    int defaultDelay = (mode.equals((SimplePostTool.DATA_MODE_WEB)) ? 10 : 0);
-    int delay = Integer.parseInt(cli.getOptionValue("delay", String.valueOf(defaultDelay)));
-    int recursive = Integer.parseInt(cli.getOptionValue("recursive", "1"));
+    int defaultDelay = (mode.equals((DATA_MODE_WEB)) ? 10 : 0);
+    delay = Integer.parseInt(cli.getOptionValue("delay", String.valueOf(defaultDelay)));
+    recursive = Integer.parseInt(cli.getOptionValue("recursive", "1"));
 
-    OutputStream out = cli.hasOption("out") ? CLIO.getOutStream() : null;
-    boolean commit = cli.hasOption("skipcommit") ? false : true;
-    boolean optimize = cli.hasOption("optimize");
+    out = cli.hasOption("out") ? CLIO.getOutStream() : null;
+    commit = cli.hasOption("skipcommit") ? false : true;
+    optimize = cli.hasOption("optimize");
 
-    String credentials = cli.getOptionValue(SolrCLI.OPTION_CREDENTIALS.getLongOpt());
+    credentials = cli.getOptionValue(SolrCLI.OPTION_CREDENTIALS.getLongOpt());
 
-    String[] args = cli.getArgs();
+    args = cli.getArgs();
 
-    SimplePostTool spt =
-        new SimplePostTool(
-            mode,
-            solrUrl,
-            credentials,
-            auto,
-            type,
-            format,
-            recursive,
-            delay,
-            fileTypes,
-            out,
-            commit,
-            optimize,
-            args);
+    execute();
+  }
 
-    spt.execute();
+  /**
+   * After initialization, call execute to start the post job. This method delegates to the correct
+   * mode method.
+   */
+  public void execute() {
+    final RTimer timer = new RTimer();
+    if (PostTool.DATA_MODE_FILES.equals(mode)) {
+      doFilesMode();
+    } else if (DATA_MODE_ARGS.equals(mode)) {
+      doArgsMode(args);
+    } else if (PostTool.DATA_MODE_WEB.equals(mode)) {
+      doWebMode();
+    } else if (DATA_MODE_STDIN.equals(mode)) {
+      doStdinMode();
+    } else {
+      return;
+    }
+
+    if (commit) {
+      commit();
+    }
+    if (optimize) {
+      optimize();
+    }
+    displayTiming((long) timer.getTime());
+  }
+
+  private void doFilesMode() {
+    currentDepth = 0;
+    // Skip posting files if special param "-" given
+    if (!args[0].equals("-")) {
+      info(
+          "Posting files to [base] url "
+              + solrUrl
+              + (!auto ? " using content-type " + (type == null ? DEFAULT_CONTENT_TYPE : type) : "")
+              + "...");
+      if (auto) {
+        info("Entering auto mode. File endings considered are " + fileTypes);
+      }
+      if (recursive > 0) {
+        info("Entering recursive mode, max depth=" + recursive + ", delay=" + delay + "s");
+      }
+      int numFilesPosted = postFiles(args, 0, out, type);
+      info(numFilesPosted + " files indexed.");
+    }
+  }
+
+  private void doArgsMode(String[] args) {
+    info("POSTing args to " + solrUrl + "...");
+    for (String a : this.args) {
+      postData(stringToStream(a), null, out, type, solrUrl);
+    }
+  }
+
+  private void doWebMode() {
+    reset();
+    int numPagesPosted = 0;
+    try {
+      if (type != null) {
+        fatal("Specifying content-type with \"-Ddata=web\" is not supported");
+      }
+      if (args[0].equals("-")) {
+        // Skip posting url if special param "-" given
+        return;
+      }
+      // Set Extracting handler as default
+      solrUrl = appendUrlPath(solrUrl, "/extract");
+
+      info("Posting web pages to Solr url " + solrUrl);
+      auto = true;
+      info(
+          "Entering auto mode. Indexing pages with content-types corresponding to file endings "
+              + fileTypes);
+      if (recursive > 0) {
+        if (recursive > MAX_WEB_DEPTH) {
+          recursive = MAX_WEB_DEPTH;
+          warn("Too large recursion depth for web mode, limiting to " + MAX_WEB_DEPTH + "...");
+        }
+        if (delay < DEFAULT_WEB_DELAY) {
+          warn(
+              "Never crawl an external web site faster than every 10 seconds, your IP will probably be blocked");
+        }
+        info("Entering recursive mode, depth=" + recursive + ", delay=" + delay + "s");
+      }
+      numPagesPosted = postWebPages(args, 0, out);
+      info(numPagesPosted + " web pages indexed.");
+    } catch (MalformedURLException e) {
+      fatal("Wrong URL trying to append /extract to " + solrUrl);
+    }
+  }
+
+  private void doStdinMode() {
+    info("POSTing stdin to " + solrUrl + "...");
+    postData(System.in, null, out, type, solrUrl);
+  }
+
+  private void reset() {
+    backlog = new ArrayList<>();
+    visited = new HashSet<>();
+  }
+
+  /**
+   * Pretty prints the number of milliseconds taken to post the content to Solr
+   *
+   * @param millis the time in milliseconds
+   */
+  private void displayTiming(long millis) {
+    SimpleDateFormat df = new SimpleDateFormat("H:mm:ss.SSS", Locale.getDefault());
+    df.setTimeZone(TimeZone.getTimeZone("UTC"));
+    CLIO.out("Time spent: " + df.format(new Date(millis)));
+  }
+
+  private boolean checkIsValidPath(File srcFile) {
+    try {
+      srcFile.toPath();
+      return true;
+    } catch (InvalidPathException e) {
+      return false;
+    }
+  }
+
+  /**
+   * Post all filenames provided in args
+   *
+   * @param args array of file names
+   * @param startIndexInArgs offset to start
+   * @param out output stream to post data to
+   * @param type default content-type to use when posting (may be overridden in auto mode)
+   * @return number of files posted
+   */
+  public int postFiles(String[] args, int startIndexInArgs, OutputStream out, String type) {
+    reset();
+    int filesPosted = 0;
+    for (int j = startIndexInArgs; j < args.length; j++) {
+      File srcFile = new File(args[j]);
+      filesPosted = getFilesPosted(out, type, srcFile);
+    }
+    return filesPosted;
+  }
+
+  /**
+   * Post all filenames provided in args
+   *
+   * @param files array of Files
+   * @param startIndexInArgs offset to start
+   * @param out output stream to post data to
+   * @param type default content-type to use when posting (may be overridden in auto mode)
+   * @return number of files posted
+   */
+  public int postFiles(File[] files, int startIndexInArgs, OutputStream out, String type) {
+    reset();
+    int filesPosted = 0;
+    for (File srcFile : files) {
+      filesPosted = getFilesPosted(out, type, srcFile);
+    }
+    return filesPosted;
+  }
+
+  private int getFilesPosted(final OutputStream out, final String type, final File srcFile) {
+    int filesPosted = 0;
+    boolean isValidPath = checkIsValidPath(srcFile);
+    if (isValidPath && srcFile.isDirectory() && srcFile.canRead()) {
+      filesPosted += postDirectory(srcFile, out, type);
+    } else if (isValidPath && srcFile.isFile() && srcFile.canRead()) {
+      filesPosted += postFiles(new File[] {srcFile}, out, type);
+    } else {
+      filesPosted += handleGlob(srcFile, out, type);
+    }
+    return filesPosted;
+  }
+
+  /**
+   * Posts a whole directory
+   *
+   * @return number of files posted total
+   */
+  private int postDirectory(File dir, OutputStream out, String type) {
+    if (dir.isHidden() && !dir.getName().equals(".")) {
+      return (0);
+    }
+    info(
+        "Indexing directory "
+            + dir.getPath()
+            + " ("
+            + dir.listFiles(fileFilter).length
+            + " files, depth="
+            + currentDepth
+            + ")");
+    int posted = 0;
+    posted += postFiles(dir.listFiles(fileFilter), out, type);
+    if (recursive > currentDepth) {
+      for (File d : dir.listFiles()) {
+        if (d.isDirectory()) {
+          currentDepth++;
+          posted += postDirectory(d, out, type);
+          currentDepth--;
+        }
+      }
+    }
+    return posted;
+  }
+
+  /**
+   * Posts a list of file names
+   *
+   * @return number of files posted
+   */
+  int postFiles(File[] files, OutputStream out, String type) {
+    int filesPosted = 0;
+    for (File srcFile : files) {
+      try {
+        if (!srcFile.isFile() || srcFile.isHidden()) {
+          continue;
+        }
+        postFile(srcFile, out, type);
+        Thread.sleep(delay * 1000L);
+        filesPosted++;
+      } catch (InterruptedException e) {
+        throw new RuntimeException(e);
+      }
+    }
+    return filesPosted;
+  }
+
+  /**
+   * This only handles file globs not full path globbing.
+   *
+   * @param globFile file holding glob path
+   * @param out outputStream to write results to
+   * @param type default content-type to use when posting (may be overridden in auto mode)
+   * @return number of files posted
+   */
+  int handleGlob(File globFile, OutputStream out, String type) {
+    int filesPosted = 0;
+    File parent = globFile.getParentFile();
+    if (parent == null) {
+      parent = new File(".");
+    }
+    String fileGlob = globFile.getName();
+    PostTool.GlobFileFilter ff = new PostTool.GlobFileFilter(fileGlob, false);
+    File[] fileList = parent.listFiles(ff);
+    if (fileList == null || fileList.length == 0) {
+      warn("No files or directories matching " + globFile);
+    } else {
+      filesPosted = postFiles(fileList, out, type);
+    }
+    return filesPosted;
+  }
+
+  /**
+   * This method takes as input a list of start URL strings for crawling, converts the URL strings
+   * to URI strings and adds each one to a backlog and then starts crawling
+   *
+   * @param args the raw input args from main()
+   * @param startIndexInArgs offset for where to start
+   * @param out outputStream to write results to
+   * @return the number of web pages posted
+   */
+  public int postWebPages(String[] args, int startIndexInArgs, OutputStream out) {
+    reset();
+    LinkedHashSet<URI> s = new LinkedHashSet<>();
+    for (int j = startIndexInArgs; j < args.length; j++) {
+      try {
+        URI uri = new URI(normalizeUrlEnding(args[j]));
+        s.add(uri);
+      } catch (URISyntaxException e) {
+        warn("Skipping malformed input URL: " + args[j]);
+      }
+    }
+    // Add URIs to level 0 of the backlog and start recursive crawling
+    backlog.add(s);
+    return webCrawl(0, out);
+  }
+
+  /**
+   * Normalizes a URL string by removing anchor part and trailing slash
+   *
+   * @return the normalized URL string
+   */
+  protected static String normalizeUrlEnding(String link) {
+    if (link.contains("#")) {
+      link = link.substring(0, link.indexOf('#'));
+    }
+    if (link.endsWith("?")) {
+      link = link.substring(0, link.length() - 1);
+    }
+    if (link.endsWith("/")) {
+      link = link.substring(0, link.length() - 1);
+    }
+    return link;
+  }
+
+  /**
+   * A very simple crawler, pulling URLs to fetch from a backlog and then recurses N levels deep if
+   * recursive&gt;0. Links are parsed from HTML through first getting an XHTML version using
+   * SolrCell with extractOnly, and followed if they are local. The crawler pauses for a default
+   * delay of 10 seconds between each fetch, this can be configured in the delay variable. This is
+   * only meant for test purposes, as it does not respect robots or anything else fancy :)
+   *
+   * @param level which level to crawl
+   * @param out output stream to write to
+   * @return number of pages crawled on this level and below
+   */
+  protected int webCrawl(int level, OutputStream out) {
+    int numPages = 0;
+    LinkedHashSet<URI> stack = backlog.get(level);
+    int rawStackSize = stack.size();
+    stack.removeAll(visited);
+    int stackSize = stack.size();
+    LinkedHashSet<URI> subStack = new LinkedHashSet<>();
+    info(
+        "Entering crawl at level "
+            + level
+            + " ("
+            + rawStackSize
+            + " links total, "
+            + stackSize
+            + " new)");
+    for (URI uri : stack) {
+      try {
+        visited.add(uri);
+        URL url = uri.toURL();
+        PostTool.PageFetcherResult result = pageFetcher.readPageFromUrl(url);
+        if (result.httpStatus == 200) {
+          url = (result.redirectUrl != null) ? result.redirectUrl : url;
+          URL postUrl =
+              new URL(
+                  appendParam(
+                      solrUrl.toString(),
+                      "literal.id="
+                          + URLEncoder.encode(url.toString(), UTF_8)
+                          + "&literal.url="
+                          + URLEncoder.encode(url.toString(), UTF_8)));
+          ByteBuffer content = result.content;
+          boolean success =
+              postData(
+                  new ByteArrayInputStream(content.array(), content.arrayOffset(), content.limit()),
+                  null,
+                  out,
+                  result.contentType,
+                  postUrl);
+          if (success) {
+            info("POSTed web resource " + url + " (depth: " + level + ")");
+            Thread.sleep(delay * 1000L);
+            numPages++;
+            // Pull links from HTML pages only
+            if (recursive > level && result.contentType.equals("text/html")) {
+              Set<URI> children =
+                  pageFetcher.getLinksFromWebPage(
+                      url,
+                      new ByteArrayInputStream(
+                          content.array(), content.arrayOffset(), content.limit()),
+                      result.contentType,
+                      postUrl);
+              subStack.addAll(children);
+            }
+          } else {
+            warn("An error occurred while posting " + uri);
+          }
+        } else {
+          warn("The URL " + uri + " returned a HTTP result status of " + result.httpStatus);
+        }
+      } catch (IOException | URISyntaxException e) {
+        warn("Caught exception when trying to open connection to " + uri + ": " + e.getMessage());
+      } catch (InterruptedException e) {
+        throw new RuntimeException(e);
+      }
+    }
+    if (!subStack.isEmpty()) {
+      backlog.add(subStack);
+      numPages += webCrawl(level + 1, out);
+    }
+    return numPages;
+  }
+
+  /**
+   * Computes the full URL based on a base url and a possibly relative link found in the href param
+   * of an HTML anchor.
+   *
+   * @param baseUrl the base url from where the link was found
+   * @param link the absolute or relative link
+   * @return the string version of the full URL
+   */
+  protected String computeFullUrl(URL baseUrl, String link) {
+    if (link == null || link.length() == 0) {
+      return null;
+    }
+    if (!link.startsWith("http")) {
+      if (link.startsWith("/")) {
+        link = baseUrl.getProtocol() + "://" + baseUrl.getAuthority() + link;
+      } else {
+        if (link.contains(":")) {
+          return null; // Skip non-relative URLs
+        }
+        String path = baseUrl.getPath();
+        if (!path.endsWith("/")) {
+          int sep = path.lastIndexOf('/');
+          String file = path.substring(sep + 1);
+          if (file.contains(".") || file.contains("?")) {
+            path = path.substring(0, sep);
+          }
+        }
+        link = baseUrl.getProtocol() + "://" + baseUrl.getAuthority() + path + "/" + link;
+      }
+    }
+    link = normalizeUrlEnding(link);
+    String l = link.toLowerCase(Locale.ROOT);
+    // Simple brute force skip images
+    if (l.endsWith(".jpg") || l.endsWith(".jpeg") || l.endsWith(".png") || l.endsWith(".gif")) {
+      return null; // Skip images
+    }
+    return link;
+  }
+
+  /**
+   * Uses the mime-type map to reverse lookup whether the file ending for our type is supported by
+   * the fileTypes option
+   *
+   * @param type what content-type to lookup
+   * @return true if this is a supported content type
+   */
+  protected boolean typeSupported(String type) {
+    for (Map.Entry<String, String> entry : mimeMap.entrySet()) {
+      if (entry.getValue().equals(type)) {
+        if (fileTypes.contains(entry.getKey())) {
+          return true;
+        }
+      }
+    }
+    return false;
+  }
+
+  /**
+   * Tests if a string is either "true", "on", "yes" or "1"
+   *
+   * @param property the string to test
+   * @return true if "on"
+   */
+  protected static boolean isOn(String property) {
+    return ("true,on,yes,1".contains(property));
+  }
+
+  static void warn(String msg) {
+    CLIO.err("PostTool: WARNING: " + msg);
+  }
+
+  static void info(String msg) {
+    CLIO.out(msg);
+  }
+
+  static void fatal(String msg) {
+    CLIO.err("PostTool: FATAL: " + msg);
+    System.exit(2);
+  }
+
+  /** Does a simple commit operation */
+  public void commit() {
+    info("COMMITting Solr index changes to " + solrUrl + "...");
+    doGet(appendParam(solrUrl.toString(), "commit=true"));
+  }
+
+  /** Does a simple optimize operation */
+  public void optimize() {
+    info("Performing an OPTIMIZE to " + solrUrl + "...");
+    doGet(appendParam(solrUrl.toString(), "optimize=true"));
+  }
+
+  /**
+   * Appends a URL query parameter to a URL
+   *
+   * @param url the original URL
+   * @param param the parameter(s) to append, separated by "&amp;"
+   * @return the string version of the resulting URL
+   */
+  public static String appendParam(String url, String param) {
+    String[] pa = param.split("&");
+    for (String p : pa) {
+      if (p.trim().length() == 0) {
+        continue;
+      }
+      String[] kv = p.split("=");
+      if (kv.length == 2) {
+        url = url + (url.contains("?") ? "&" : "?") + kv[0] + "=" + kv[1];
+      } else {
+        warn("Skipping param " + p + " which is not on form key=value");
+      }
+    }
+    return url;
+  }
+
+  /** Opens the file and posts its contents to the solrUrl, writes to response to output. */
+  public void postFile(File file, OutputStream output, String type) {
+    InputStream is = null;
+    try {
+      URL url = solrUrl;
+      String suffix = "";
+      if (auto) {
+        if (type == null) {
+          type = guessType(file);
+        }
+        // TODO: Add a flag that disables /update and sends all to /update/extract, to avoid CSV,
+        // JSON, and XML files
+        // TODO: from being interpreted as Solr documents internally
+        if (type.equals("application/json") && !PostTool.FORMAT_SOLR.equals(format)) {
+          suffix = "/json/docs";
+          String urlStr = appendUrlPath(solrUrl, suffix).toString();
+          url = new URL(urlStr);
+        } else if (type.equals("application/xml")
+            || type.equals("text/csv")
+            || type.equals("application/json")) {
+          // Default handler
+        } else {
+          // SolrCell
+          suffix = "/extract";
+          String urlStr = appendUrlPath(solrUrl, suffix).toString();
+          if (!urlStr.contains("resource.name")) {
+            urlStr =
+                appendParam(
+                    urlStr, "resource.name=" + URLEncoder.encode(file.getAbsolutePath(), UTF_8));
+          }
+          if (!urlStr.contains("literal.id")) {
+            urlStr =
+                appendParam(
+                    urlStr, "literal.id=" + URLEncoder.encode(file.getAbsolutePath(), UTF_8));
+          }
+          url = new URL(urlStr);
+        }
+      } else {
+        if (type == null) {
+          type = DEFAULT_CONTENT_TYPE;
+        }
+      }
+      info(
+          "POSTing file "
+              + file.getName()
+              + (auto ? " (" + type + ")" : "")
+              + " to [base]"
+              + suffix
+              + (mockMode ? " MOCK!" : ""));
+      is = new FileInputStream(file);
+      postData(is, file.length(), output, type, url);
+    } catch (IOException e) {
+      warn("Can't open/read file: " + file);
+    } finally {
+      try {
+        if (is != null) {
+          is.close();
+        }
+      } catch (IOException e) {
+        fatal("IOException while closing file: " + e);
+      }
+    }
+  }
+
+  /**
+   * Appends to the path of the URL
+   *
+   * @param url the URL
+   * @param append the path to append
+   * @return the final URL version
+   */
+  protected static URL appendUrlPath(URL url, String append) throws MalformedURLException {
+    return new URL(
+        url.getProtocol()
+            + "://"
+            + url.getAuthority()
+            + url.getPath()
+            + append
+            + (url.getQuery() != null ? "?" + url.getQuery() : ""));
+  }
+
+  /**
+   * Guesses the type of file, based on file name suffix Returns "application/octet-stream" if no
+   * corresponding mimeMap type.
+   *
+   * @param file the file
+   * @return the content-type guessed
+   */
+  protected static String guessType(File file) {
+    String name = file.getName();
+    String suffix = name.substring(name.lastIndexOf('.') + 1);
+    String type = mimeMap.get(suffix.toLowerCase(Locale.ROOT));
+    return (type != null) ? type : "application/octet-stream";
+  }
+
+  /** Performs a simple get on the given URL */
+  public void doGet(String url) {
+    try {
+      doGet(new URL(url));
+    } catch (MalformedURLException e) {
+      warn("The specified URL " + url + " is not a valid URL. Please check");
+    }
+  }
+
+  /** Performs a simple get on the given URL */
+  public void doGet(URL url) {
+    try {
+      if (mockMode) {
+        return;
+      }
+      HttpURLConnection urlc = (HttpURLConnection) url.openConnection();
+      basicAuth(urlc);
+      urlc.connect();
+      checkResponseCode(urlc);
+    } catch (IOException e) {
+      warn("An error occurred getting data from " + url + ". Please check that Solr is running.");
+    } catch (Exception e) {
+      warn("An error occurred getting data from " + url + ". Message: " + e.getMessage());
+    }
+  }
+
+  /**
+   * Reads data from the data stream and posts it to solr, writes to the response to output
+   *
+   * @return true if success
+   */
+  public boolean postData(
+      InputStream data, Long length, OutputStream output, String type, URL url) {
+    if (mockMode) {
+      return true;
+    }
+
+    boolean success = true;
+    if (type == null) {
+      type = DEFAULT_CONTENT_TYPE;
+    }
+    HttpURLConnection urlConnection = null;
+    try {
+      try {
+        urlConnection = (HttpURLConnection) url.openConnection();
+        try {
+          urlConnection.setRequestMethod("POST");
+        } catch (ProtocolException e) {
+          fatal("Shouldn't happen: HttpURLConnection doesn't support POST??" + e);
+        }
+        urlConnection.setDoOutput(true);
+        urlConnection.setDoInput(true);
+        urlConnection.setUseCaches(false);
+        urlConnection.setAllowUserInteraction(false);
+        urlConnection.setRequestProperty("Content-type", type);
+        basicAuth(urlConnection);
+        if (null != length) {
+          urlConnection.setFixedLengthStreamingMode(length);
+        } else {
+          urlConnection.setChunkedStreamingMode(-1); // use JDK default chunkLen, 4k in Java 8.
+        }
+        urlConnection.connect();
+      } catch (IOException e) {
+        fatal("Connection error (is Solr running at " + solrUrl + " ?): " + e);
+        success = false;
+      } catch (Exception e) {
+        fatal("POST failed with error " + e.getMessage());
+      }
+
+      try (final OutputStream out = urlConnection.getOutputStream()) {
+        pipe(data, out);
+      } catch (IOException e) {
+        fatal("IOException while posting data: " + e);
+      }
+
+      try {
+        success &= checkResponseCode(urlConnection);
+        try (final InputStream in = urlConnection.getInputStream()) {
+          pipe(in, output);
+        }
+      } catch (IOException e) {
+        warn("IOException while reading response: " + e);
+        success = false;
+      } catch (GeneralSecurityException e) {
+        fatal(
+            "Looks like Solr is secured and would not let us in. Try with another user in '-u' parameter");
+      }
+    } finally {
+      if (urlConnection != null) {
+        urlConnection.disconnect();
+      }
+    }
+    return success;
+  }
+
+  private void basicAuth(HttpURLConnection urlc) throws Exception {
+    if (urlc.getURL().getUserInfo() != null) {
+      String encoding =
+          Base64.getEncoder().encodeToString(urlc.getURL().getUserInfo().getBytes(US_ASCII));
+      urlc.setRequestProperty("Authorization", "Basic " + encoding);
+    } else if (credentials != null) {
+      if (!credentials.contains(":")) {
+        throw new Exception("credentials '" + credentials + "' must be of format user:pass");
+      }
+      urlc.setRequestProperty(
+          "Authorization",
+          "Basic " + Base64.getEncoder().encodeToString(credentials.getBytes(UTF_8)));
+    }
+  }
+
+  private static boolean checkResponseCode(HttpURLConnection urlc)
+      throws IOException, GeneralSecurityException {
+    if (urlc.getResponseCode() >= 400) {
+      warn(
+          "Solr returned an error #"
+              + urlc.getResponseCode()
+              + " ("
+              + urlc.getResponseMessage()
+              + ") for url: "
+              + urlc.getURL());
+      Charset charset = StandardCharsets.ISO_8859_1;
+      final String contentType = urlc.getContentType();
+      // code cloned from ContentStreamBase, but post.jar should be standalone!
+      if (contentType != null) {
+        int idx = contentType.toLowerCase(Locale.ROOT).indexOf("charset=");
+        if (idx > 0) {
+          charset = Charset.forName(contentType.substring(idx + "charset=".length()).trim());
+        }
+      }
+      // Print the response returned by Solr
+      try (InputStream errStream = urlc.getErrorStream()) {
+        if (errStream != null) {
+          BufferedReader br = new BufferedReader(new InputStreamReader(errStream, charset));
+          final StringBuilder response = new StringBuilder("Response: ");
+          int ch;
+          while ((ch = br.read()) != -1) {
+            response.append((char) ch);
+          }
+          warn(response.toString().trim());
+        }
+      }
+      if (urlc.getResponseCode() == 401) {
+        throw new GeneralSecurityException(
+            "Solr requires authentication (response 401). Please try again with '-u' option");
+      }
+      if (urlc.getResponseCode() == 403) {
+        throw new GeneralSecurityException(
+            "You are not authorized to perform this action against Solr. (response 403)");
+      }
+      return false;
+    }
+    return true;
+  }
+
+  /**
+   * Converts a string to an input stream
+   *
+   * @param s the string
+   * @return the input stream
+   */
+  public static InputStream stringToStream(String s) {
+    return new ByteArrayInputStream(s.getBytes(StandardCharsets.UTF_8));
+  }
+
+  /**
+   * Pipes everything from the source to the dest. If dest is null, then everything is read from
+   * source and thrown away.
+   */
+  private static void pipe(InputStream source, OutputStream dest) throws IOException {
+    byte[] buf = new byte[1024];
+    int read = 0;
+    while ((read = source.read(buf)) >= 0) {
+      if (null != dest) {
+        dest.write(buf, 0, read);
+      }
+    }
+    if (null != dest) {
+      dest.flush();
+    }
+  }
+
+  public FileFilter getFileFilterFromFileTypes(String fileTypes) {
+    String glob;
+    if (fileTypes.equals("*")) {
+      glob = ".*";
+    } else {
+      glob = "^.*\\.(" + fileTypes.replace(",", "|") + ")$";
+    }
+    return new PostTool.GlobFileFilter(glob, true);
+  }
+
+  //
+  // Utility methods for XPath handing
+  //
+
+  /** Gets all nodes matching an XPath */
+  public static NodeList getNodesFromXP(Node n, String xpath) throws XPathExpressionException {
+    XPathFactory factory = XPathFactory.newInstance();
+    XPath xp = factory.newXPath();
+    XPathExpression expr = xp.compile(xpath);
+    return (NodeList) expr.evaluate(n, XPathConstants.NODESET);
+  }
+
+  /**
+   * Gets the string content of the matching an XPath
+   *
+   * @param n the node (or doc)
+   * @param xpath the xpath string
+   * @param concatAll if true, text from all matching nodes will be concatenated, else only the
+   *     first returned
+   */
+  public static String getXP(Node n, String xpath, boolean concatAll)
+      throws XPathExpressionException {
+    NodeList nodes = getNodesFromXP(n, xpath);
+    StringBuilder sb = new StringBuilder();
+    if (nodes.getLength() > 0) {
+      for (int i = 0; i < nodes.getLength(); i++) {
+        sb.append(nodes.item(i).getNodeValue()).append(' ');
+        if (!concatAll) {
+          break;
+        }
+      }
+      return sb.toString().trim();
+    } else return "";
+  }
+
+  /** Takes a string as input and returns a DOM */
+  public static Document makeDom(byte[] in)
+      throws SAXException, IOException, ParserConfigurationException {
+    InputStream is = new ByteArrayInputStream(in);
+    Document dom = DocumentBuilderFactory.newInstance().newDocumentBuilder().parse(is);
+    return dom;
+  }
+
+  /** Inner class to filter files based on glob wildcards */
+  static class GlobFileFilter implements FileFilter {
+    private String _pattern;
+    private Pattern p;
+
+    public GlobFileFilter(String pattern, boolean isRegex) {
+      _pattern = pattern;
+      if (!isRegex) {
+        _pattern =
+            _pattern
+                .replace("^", "\\^")
+                .replace("$", "\\$")
+                .replace(".", "\\.")
+                .replace("(", "\\(")
+                .replace(")", "\\)")
+                .replace("+", "\\+")
+                .replace("*", ".*")
+                .replace("?", ".");
+        _pattern = "^" + _pattern + "$";
+      }
+
+      try {
+        p = Pattern.compile(_pattern, Pattern.CASE_INSENSITIVE);
+      } catch (PatternSyntaxException e) {
+        fatal("Invalid type list " + pattern + ". " + e.getDescription());
+      }
+    }
+
+    @Override
+    public boolean accept(File file) {
+      return p.matcher(file.getName()).find();
+    }
+  }
+
+  //
+  // Simple crawler class which can fetch a page and check for robots.txt
+  //
+  class PageFetcher {
+    Map<String, List<String>> robotsCache;
+    static final String DISALLOW = "Disallow:";
+
+    public PageFetcher() {
+      robotsCache = new HashMap<>();
+    }
+
+    public PageFetcherResult readPageFromUrl(URL u) throws URISyntaxException {
+      PostTool.PageFetcherResult res = new PostTool.PageFetcherResult();
+      try {
+        if (isDisallowedByRobots(u)) {
+          warn("The URL " + u + " is disallowed by robots.txt and will not be crawled.");
+          res.httpStatus = 403;
+          URI uri = u.toURI();
+          visited.add(uri);
+          return res;
+        }
+        res.httpStatus = 404;
+        HttpURLConnection conn = (HttpURLConnection) u.openConnection();
+        conn.setRequestProperty(
+            "User-Agent",
+            "PostTool-crawler/" + SolrVersion.LATEST_STRING + " (https://solr.apache.org/)");
+        conn.setRequestProperty("Accept-Encoding", "gzip, deflate");
+        conn.connect();
+        res.httpStatus = conn.getResponseCode();
+        if (!normalizeUrlEnding(conn.getURL().toString())
+            .equals(normalizeUrlEnding(u.toString()))) {
+          info("The URL " + u + " caused a redirect to " + conn.getURL());
+          u = conn.getURL();
+          res.redirectUrl = u;
+          URI uri = u.toURI();
+          visited.add(uri);
+        }
+        if (res.httpStatus == 200) {
+          // Raw content type of form "text/html; encoding=utf-8"
+          String rawContentType = conn.getContentType();
+          String type = rawContentType.split(";")[0];
+          if (typeSupported(type) || "*".equals(fileTypes)) {
+            String encoding = conn.getContentEncoding();
+            InputStream is;
+            if (encoding != null && encoding.equalsIgnoreCase("gzip")) {
+              is = new GZIPInputStream(conn.getInputStream());
+            } else if (encoding != null && encoding.equalsIgnoreCase("deflate")) {
+              is = new InflaterInputStream(conn.getInputStream(), new Inflater(true));
+            } else {
+              is = conn.getInputStream();
+            }
+
+            // Read into memory, so that we later can pull links from the page without re-fetching
+            res.content = Utils.toByteArray(is);
+            is.close();
+          } else {
+            warn("Skipping URL with unsupported type " + type);
+            res.httpStatus = 415;
+          }
+        }
+      } catch (IOException e) {
+        warn("IOException when reading page from url " + u + ": " + e.getMessage());
+      }
+      return res;
+    }
+
+    public boolean isDisallowedByRobots(URL url) {
+      String host = url.getHost();
+      String strRobot = url.getProtocol() + "://" + host + "/robots.txt";
+      List<String> disallows = robotsCache.get(host);
+      if (disallows == null) {
+        disallows = new ArrayList<>();
+        URL urlRobot;
+        try {
+          urlRobot = new URL(strRobot);
+          disallows = parseRobotsTxt(urlRobot.openStream());
+        } catch (MalformedURLException e) {
+          return true; // We cannot trust this robots URL, should not happen
+        } catch (IOException e) {
+          // There is no robots.txt, will cache an empty disallow list
+        }
+      }
+
+      robotsCache.put(host, disallows);
+
+      String strURL = url.getFile();
+      for (String path : disallows) {
+        if (path.equals("/") || strURL.indexOf(path) == 0) return true;
+      }
+      return false;
+    }
+
+    /**
+     * Very simple robots.txt parser which obeys all Disallow lines regardless of user agent or
+     * whether there are valid Allow: lines.
+     *
+     * @param is Input stream of the robots.txt file
+     * @return a list of disallow paths
+     * @throws IOException if problems reading the stream
+     */
+    protected List<String> parseRobotsTxt(InputStream is) throws IOException {
+      List<String> disallows = new ArrayList<>();
+      BufferedReader r = new BufferedReader(new InputStreamReader(is, StandardCharsets.UTF_8));
+      String l;
+      while ((l = r.readLine()) != null) {
+        String[] arr = l.split("#");
+        if (arr.length == 0) continue;
+        l = arr[0].trim();
+        if (l.startsWith(DISALLOW)) {
+          l = l.substring(DISALLOW.length()).trim();
+          if (l.length() == 0) continue;
+          disallows.add(l);
+        }
+      }
+      is.close();
+      return disallows;
+    }
+
+    /**
+     * Finds links on a web page, using /extract?extractOnly=true
+     *
+     * @param u the URL of the web page
+     * @param is the input stream of the page
+     * @param type the content-type
+     * @param postUrl the URL (typically /solr/extract) in order to pull out links
+     * @return a set of URIs parsed from the page
+     */
+    protected Set<URI> getLinksFromWebPage(URL u, InputStream is, String type, URL postUrl) {
+      Set<URI> linksFromPage = new HashSet<>();
+      URL url = null;
+      try {
+        ByteArrayOutputStream os = new ByteArrayOutputStream();
+        URL extractUrl = new URL(appendParam(postUrl.toString(), "extractOnly=true"));
+        extractUrl = new URL(appendParam(extractUrl.toString(), "wt=xml"));
+        boolean success = postData(is, null, os, type, extractUrl);
+        if (success) {
+          Document d = makeDom(os.toByteArray());
+          String innerXml = getXP(d, "/response/str/text()[1]", false);
+          d = makeDom(innerXml.getBytes(StandardCharsets.UTF_8));
+          NodeList links = getNodesFromXP(d, "/html/body//a/@href");
+          for (int i = 0; i < links.getLength(); i++) {
+            String link = links.item(i).getTextContent();
+            link = computeFullUrl(u, link);
+            if (link == null) {
+              continue;
+            }
+            URI newUri = new URI(link);
+            if (newUri.getAuthority() == null || !newUri.getAuthority().equals(u.getAuthority())) {
+              linksFromPage.add(newUri);
+            }
+          }
+        }
+      } catch (MalformedURLException e) {
+        warn("Malformed URL " + url);
+      } catch (IOException e) {
+        warn("IOException opening URL " + url + ": " + e.getMessage());
+      } catch (Exception e) {
+        throw new RuntimeException(e);
+      }
+
+      return linksFromPage;
+    }
+  }
+
+  /** Utility class to hold the result form a page fetch */
+  public static class PageFetcherResult {
+    int httpStatus = 200;
+    String contentType = "text/html";
+    URL redirectUrl = null;
+    ByteBuffer content;
   }
 }
